Houston, we have data! (from a use case, even!)
After generating a YAML file with fixture data (50 articles, with
randomly-generated data including a smaller set of randomly-generated
author names), we set up Wisper message traffic to query The World for
the "current user", and the list of all `Article`s which that user may
view.

This is not likely to be the use case's final form (the name, "summarise
content", should tip the reader off that there is likely more "content"
than simply a list of articles), but it "works" as demonstrating the
Wisper message traffic between the use case and an "authorisation port"
that "knows" what the "current user" is, and the "`Article` persistence
port", that can supply a list of all `Article` entities which that user
may view. **IMPORTANT NOTE:** There is no protocol yet defined for the
article-persistence port to actually do the filtering; in particular,
there is no linkage between the persistence provider and any sort of
authorisation that would do (or provide information necessary for) that
filtering.

We're again bumping up against the limitations of Wisper as a remote
data *provider* in a query/response fashion; with subscriptions limited
to either individual object pairs or to an object listening for "global"
broadcast messages (however scoped/filtered), this *feels* very fragile.
Don't get us wrong; we *love* Wisper in its designed-for use case, which
is basically the Observer pattern of *unidirectional* notification. It
wobbles considerably more the farther outside that role it's asked to go.

4 tests, 4 assertions, 0 failures, 0 errors, 0 skips
Coverage: suspended; see prolog_core Issue 1
RuboCop: 6 files inspected, no offenses detected
Flay: Total score 0
Flog: Total 10.8; method average 2.2; max 3.9 (Prolog::UseCases::SummariseContent#load_article_list)
Reek: 0 warnings