Broke out `ProposeEditContribution` form object.
Also note the addition of a `ProposeEditContributionTestEnhancer` class,
which (initially) simply adds the ability to read or write the Markdown-
converter object in the form object from outside the class. This enables
us to stick a new one in that, for instance, counts the number of times
that a method on it is called. We probably don't want that on the code
that goes into production, do we? :grinning:

171 tests, 194 assertions, 0 failures, 0 errors, 20 skips
Coverage: suspended; see prolog_core Issue 1
RuboCop: 52 files inspected, no offenses detected
Flay: Total score 0
Flog: Total 426.8; method average 3.6; max 8.1 (Prolog::UseCases::SummariseContent::KeywordCounter#list_keywords_by_count)
Reek: 0 warnings