Beginning of `NodeCleanup::Element`.
For the first time, we've managed to wrap our heads around what we hope
(and believe) is the complete set of requirements for whitespace in an
element. Significantly, we draw a distinction between spacing of *block*
elements and *inline* elements. Quoting the internal doc for the new
`MarkdownToHtml::Dumper::NodeCleanup::Element` class (reformatted):

> Top-level class to clean a single HTML element string (possibly
> including child elements) so that its textual representation
> complies with our whitespae conventions. In summary, these are:
>
> 1. Any sequence of whitespace characters within a string will be
>    replaced by a single space;
> 2. Any string which is the *first* child node of an element will
>    have any leading whitespace removed;
> 3. Any string which is the first child node of an element will have
>    Any trailing whitespace replaced by a single space;
> 4. Any string which is the *last* child node of an element will have
>    any trailing whitespace removed;
> 5. Any string which is the last child node of an element will have
>    any leading whitespace replaced by a single space;
> 6. Any string which is an intermediating child node of an element,
>    i.e., which occurs between one child element of the parent
>    element and another child element, will have any leading *or*
>    trailing whitespace replaced by a single space.
> 7. Any whitespace between *block level elements* (such as between
>    two `p` elements or two `div` elements, but *not* between two
>    `span` elements or between an `em` and `span` elements) will be
>    removed.
>
> It Would Be Very Nice If Ox (preferably) or Nokogiri could *tell* us
> if an element is a block element or not, but we're on our own.

That bit in Item 7 is the epiphany we've been beating our heads against
various surfaces for three days trying to force to the surface. This also
suggests that the list of classes in the "I Can Name That Tune" comment
on Issue #23 will need further tweaking, as neither of the concepts
needed to implement Item 7 yet exist in code (as alluded to in the last
two lines of the quote.)

166 tests, 200 assertions, 0 failures, 0 errors, 0 skips
Coverage: suspended; see prolog_core Issue 1
RuboCop: 55 files inspected, no offenses detected
Flay: Total score 0
Flog: Total 448.7; method average 3.6; max 8.1 (Prolog::UseCases::SummariseContent::KeywordCounter#list_keywords_by_count)
Reek: 0 warnings