Added `pandoc-ruby` to Gemspec; basic functionality test added.
[Pandoc](http://pandoc.org) is a "universal document converter"; we're
interested in it primarily for converting HTML to Markdown. "Hang on,"
you might be thinking, "don't we already have a Markdown-to-HTML
converter?" Yes, we do, but if we had it to do over again, we probably
wouldn't. HTML-to-Markdown conversion has a couple of uses; it *should*
be a bit easier for us to wind up with valid, well-formed HTML after a
simple contribution if the Markdown content of the replacement is
actually used to replace the *Markdown content being replaced.*

Consider the following markup as rendered within an article body
(reformatted for readability; the actual source would be on a single line
without newlines or excess inter-element whitespace):

```html
<ul>
  <li>A first item;</li>
  <li>A <em>second</em> item; and</li>
  <li>A final item.</li>
</ul>
```

Now, suppose that a member wished to change `<em>second</em>` to
`second`. If we naively convert the proposed replacement content from
Markdown to HTML, we'd wind up with malformed HTML once the Proposal was
Accepted:

```html
<ul>
  <li>A first item;</li>
  <li>A <a id="whatever-begin"></a><p>second</p><a id="whatever-end"></a> item; and</li>
  <li>A final item.</li>
</ul>
```

Any strictly conforming parser, including the Ox which we use pervasively
in our existing code, would choke on the embedded `<p>` within an `<li>`.

But if the replacement is made to a *Markdown representation* of the
article content:

```markdown
-   A first item;
-   A <a id="whatever-begin"></a>*second*<span id="whatever-end"></span> item; and
-   A final item.
```

then replacing `*second*` with `second` (and wrapping the markers around
it) would do no harm, and the HTML converted from the updated Markdown
would Do The Right Thing.

**IMPORTANT NOTE:** During the contrivance of this example, it was
discovered that *no* Markdown-to-HTML converter, either our existing one
or Pandoc, would convert the example list items properly if an HTML tag
(such as `<a>` or `<span>`) was the *first text* of a list item. For
example, the following would have Problems:

```markdown
-   A first item;
-   <a id="whatever-begin"></a>A *second*<span id="whatever-end"></span> item; and
-   A final item.
```

The converted HTML would have *two* list items, and the `-` at the start
of the second line would be renderd as a literal character. Oops.

230 tests, 277 assertions, 0 failures, 0 errors, 0 skips
Coverage: suspended; see prolog_core Issue 1
RuboCop: 59 files inspected, no offenses detected
Flay: Total score 0
Flog: Total 522.3; method average 4.0; max 8.1 (Prolog::UseCases::SummariseContent::KeywordCounter#list_keywords_by_count)
Reek: 0 warnings