Removed Entities::Proposal, leaving EditContribution::Proposed.
But we're not ultimately happy with this, for several reasons. We think
that the attribute naming is cleaner in the removed class (`proposer` as
opposed to `user_name`, for instance), but there's also the question of
how these entities are to be *organised.*

The class we have now is `Prolog::Entities::EditContribution::Proposed`,
which implies that a proposal that has been Responded to will differ.
Fine; it likely should, and the proposal for an Edit Contribution will be
somewhat different than that for a Challenge Contribution or any other
"type" of Contribution. But by having `EditContribution` as a standalone,
top-level entity class, we're implying that its *altogether* different
than any other type of Contribution, and that's certainly *not* the case.
Any Contribution will have a Proposer; any Contribution will be tied to a
specific Article, generally to a specific fragment of content *within*
that Article, and so on. Whatever mechanism we use to DRY up repeated
code within broadly-similar classes (straight-up inheritance, mixins,
etc), the naming conventions we've stumbled into (with either of the
pre-existing classes, actually) aren't likely to be helpful in the long
run, and smoothing the roughest of edges now is a relatively cheap
operation.

Some further cleanup remains to be done.

296 tests, 350 assertions, 0 failures, 0 errors, 0 skips
Coverage: 2481 / 2483 LOC (99.92%) covered.
RuboCop: 70 files inspected, no offenses detected
Flay: Total score 0
Flog: Total 691.8; method average 3.6; max 5.7 (Prolog::UseCases::ProposeEditContribution::FormObject#wrap_contribution_with)
Reek: 0 warnings