All specified preconditions now verified.
The use-case class isn't pretty (as Reek attests without silencing
pragmas), but it works. If the preconditions are properly met, the
passed-in responder is called, passing in the proposal. This *should* be
everything we know that the UI delivery mechanism needs from us to
present the proposal to the logged-in Author and accept a response,
which is passed back to us. That response *must* be either of the
symbols `:accepted` or `:rejected`.

Closes #40, with the caveat mentioned in [this comment](https://github.com/TheProlog/prolog-use_cases/issues/40#issuecomment-220314901)
on that issue.

312 tests, 358 assertions, 0 failures, 0 errors, 0 skips
Coverage: 2804 / 2805 LOC (99.96%) covered.
RuboCop: 83 files inspected, no offenses detected
Flay: Total score 0
Flog: Total 873.1; method average 3.4; max 5.7 (Prolog::UseCases::RegisterNewMember::FormObject#unique_error_messages)
Reek: 0 warnings