Added `:article` attribute to `ProposeEditContribution::Result`.
*Important: Earlier drafts of the Wiki spec and the code spec specified
that, in the case of a Contribution Proposed by the Author of an Article,
this use case was to "automatically" Accept the Contribution, using logic
essentially identical to half of the "Accept or Reject Proposed Edit
Contribu­tion" use case. This had the effect of, finally, rubbing our
noses in the fact that "Accept or Reject" is a Frankenstein object: the
unnatur­al union of multiple things that are best done or used in other
ways, depending on an artifice (the injected author response) to
indicate which of multiple personalities should be exhibited. That's not
the sort of thing most people think of as "well-designed software."
Rath­er, this should be a single-path action that does just what it says;
it's the responsibility of Some Out­side Code to say "hey, the current
user is the article author; we need to run the Approve Proposed
Contribution use case directly after successfully running the Propose
Edit Contribution UC".

296 tests, 343 assertions, 0 failures, 0 errors, 0 skips
Coverage: 2742 / 2748 LOC (99.78%) covered.
RuboCop: 84 files inspected, no offenses detected
Flay: Total score 0
Flog: Total 868.8; method average 3.4; max 5.7 (Prolog::UseCases::RegisterNewMember::FormObject#unique_error_messages)
Reek: 0 warnings