Added simple article-verification specs.
If the same Author name and Title are on the Article going in as on the
Article in the result, we're more certain that we are in fact dealing
with two entities representing the same article.

298 tests, 345 assertions, 0 failures, 0 errors, 0 skips
Coverage: 2746 / 2752 LOC (99.78%) covered.
RuboCop: 84 files inspected, no offenses detected
Flay: Total score 0
Flog: Total 868.8; method average 3.4; max 5.7 (Prolog::UseCases::RegisterNewMember::FormObject#unique_error_messages)
Reek: 0 warnings